管理器检测花园是否完全启动
