+++ lib/iirose-universal-sdk/net/ws.ts
@@ -5,11 +5,13 @@ export const events = new EventEmitter()
 
 getHookedSocket(
   (data) => {
-    events.emit('rx', data)
+    // 收到信息
+    // events.emit('rx', data)
     return true
   },
   (data) => {
-    events.emit('tx', data)
+    // 发送信息
+    // events.emit('tx', data)
     return true
   }
 )
