+++ rollup.config.js
@@ -2,6 +2,7 @@ import resolve from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import typescript from "rollup-plugin-typescript2";
 import nodePolyfills from 'rollup-plugin-polyfill-node';
+import server from 'rollup-plugin-serve';
 import { readdirSync } from 'fs';
 import path from 'path';
 
@@ -22,7 +23,11 @@ export default files.map(file => ({
     resolve(), // 解析第三方模块
     commonjs(), // 将 CommonJS 模块转换为 ES6 模块
     typescript({ tsconfig: './tsconfig.json', tsconfigOverride: override }),
-    nodePolyfills()
+    nodePolyfills(),
+    server({
+      contentBase: [outputDir],
+      port: 8080
+    })
   ]
 }));
